chore: use operations and models subfolders
