+++ packages/js/src/plugins/systemModule/index.ts
@@ -1,5 +1,4 @@
-export * from './createAccount';
+export * from './operations';
 export * from './plugin';
 export * from './SystemBuildersClient';
 export * from './SystemClient';
-export * from './transferSol';
