+++ packages/js/src/plugins/systemModule/operations/transferSol.ts
@@ -9,7 +9,7 @@ import {
   useOperation,
 } from '@/types';
 import { DisposableScope, TransactionBuilder } from '@/utils';
-import { SendAndConfirmTransactionResponse } from '../rpcModule';
+import { SendAndConfirmTransactionResponse } from '../../rpcModule';
 
 // -----------------
 // Operation
