+++ packages/js/src/plugins/tokenModule/index.ts
@@ -1,24 +1,11 @@
 export * from './accounts';
-export * from './approveTokenDelegateAuthority';
 export * from './constants';
-export * from './createMint';
-export * from './createToken';
-export * from './createTokenWithMint';
 export * from './errors';
-export * from './findMintByAddress';
-export * from './findTokenByAddress';
-export * from './findTokenWithMintByAddress';
-export * from './findTokenWithMintByMint';
-export * from './freezeTokens';
 export * from './gpaBuilders';
-export * from './Mint';
-export * from './mintTokens';
+export * from './models';
+export * from './operations';
 export * from './pdas';
 export * from './plugin';
 export * from './program';
-export * from './revokeTokenDelegateAuthority';
-export * from './sendTokens';
-export * from './thawTokens';
-export * from './Token';
 export * from './TokenBuildersClient';
 export * from './TokenClient';
