+++ packages/js/src/plugins/tokenModule/models/Mint.ts
@@ -1,8 +1,8 @@
 import type { PublicKey } from '@solana/web3.js';
 import { amount, SplTokenCurrency, SplTokenAmount } from '@/types';
 import { assert, Option } from '@/utils';
-import { MintAccount } from './accounts';
-import { WRAPPED_SOL_MINT } from './constants';
+import { MintAccount } from '../accounts';
+import { WRAPPED_SOL_MINT } from '../constants';
 
 /** @group Models */
 export type Mint = {
