+++ packages/js/src/plugins/tokenModule/models/index.ts
@@ -1,0 +1,2 @@
+export * from './Mint';
+export * from './Token';
