+++ packages/js/src/plugins/tokenModule/operations/approveTokenDelegateAuthority.ts
@@ -12,9 +12,9 @@ import {
 import { TransactionBuilder } from '@/utils';
 import { createApproveInstruction } from '@solana/spl-token';
 import { ConfirmOptions, PublicKey } from '@solana/web3.js';
-import { SendAndConfirmTransactionResponse } from '../rpcModule';
-import { findAssociatedTokenAccountPda } from './pdas';
-import { TokenProgram } from './program';
+import { SendAndConfirmTransactionResponse } from '../../rpcModule';
+import { findAssociatedTokenAccountPda } from '../pdas';
+import { TokenProgram } from '../program';
 
 // -----------------
 // Operation
