+++ packages/js/src/plugins/tokenModule/operations/findTokenByAddress.ts
@@ -1,8 +1,8 @@
 import { Metaplex } from '@/Metaplex';
 import { Operation, OperationHandler, useOperation } from '@/types';
 import type { Commitment, PublicKey } from '@solana/web3.js';
-import { toTokenAccount } from './accounts';
-import { Token, toToken } from './Token';
+import { toTokenAccount } from '../accounts';
+import { Token, toToken } from '../models/Token';
 
 // -----------------
 // Operation
