+++ packages/js/src/plugins/tokenModule/operations/findTokenWithMintByMint.ts
@@ -1,11 +1,11 @@
 import { Metaplex } from '@/Metaplex';
 import { Operation, OperationHandler, useOperation } from '@/types';
 import type { Commitment, PublicKey } from '@solana/web3.js';
-import { toMintAccount, toTokenAccount } from './accounts';
-import { TokenAndMintDoNotMatchError } from './errors';
-import { toMint } from './Mint';
-import { findAssociatedTokenAccountPda } from './pdas';
-import { TokenWithMint, toTokenWithMint } from './Token';
+import { toMintAccount, toTokenAccount } from '../accounts';
+import { TokenAndMintDoNotMatchError } from '../errors';
+import { toMint } from '../models/Mint';
+import { findAssociatedTokenAccountPda } from '../pdas';
+import { TokenWithMint, toTokenWithMint } from '../models/Token';
 
 // -----------------
 // Operation
